@@ -1,22 +1,21 @@
 package com.gafahtec.consultorio.validators;
 
 import com.gafahtec.consultorio.annotations.UniqueEmail;
-import com.gafahtec.consultorio.model.Usuario;
 import com.gafahtec.consultorio.repository.IUsuarioRepository;
-import lombok.AllArgsConstructor;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
 public class UniqueEmailValidator implements ConstraintValidator<UniqueEmail, String> {
 
 
-    IUsuarioRepository iUsuarioRepository;
+	IUsuarioRepository iUsuarioRepository;
 
     @Override
     public boolean isValid(String value, ConstraintValidatorContext context) {
-        Usuario usuario = iUsuarioRepository.findByEmail(value);
+        var usuario = iUsuarioRepository.findByEmail(value);
         if (usuario == null) {
             return true;
         }
