@@ -1,13 +1,14 @@
 package com.gafahtec.consultorio.annotations;
 
-import com.gafahtec.consultorio.validators.UniqueEmailValidator;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
+import com.gafahtec.consultorio.validators.UniqueEmailValidator;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 
 @Constraint(validatedBy = UniqueEmailValidator.class)
 @Target(ElementType.FIELD)
