@@ -1,14 +1,10 @@
 package com.gafahtec.consultorio.controller;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.validation.Valid;
+import java.util.Set;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -22,13 +18,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.gafahtec.consultorio.dto.request.CitaRequest;
+import com.gafahtec.consultorio.dto.response.CitaResponse;
 import com.gafahtec.consultorio.exception.ResourceNotFoundException;
-import com.gafahtec.consultorio.model.Cita;
-import com.gafahtec.consultorio.model.Cliente;
-import com.gafahtec.consultorio.model.Horario;
-import com.gafahtec.consultorio.model.ProgramacionDetalle;
 import com.gafahtec.consultorio.service.ICitaService;
 
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 
 @RestController
@@ -39,81 +33,71 @@ public class CitaController {
     private ICitaService iCitaService;
 
     @GetMapping
-    public ResponseEntity<List<Cita>> listar() throws ResourceNotFoundException {
-        List<Cita> lista = iCitaService.listar();
+    public ResponseEntity<Set<CitaResponse>> listar() throws ResourceNotFoundException {
+    	var lista = iCitaService.listar();
         if (lista.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-        return new ResponseEntity<List<Cita>>(lista, HttpStatus.OK);
+        return new ResponseEntity<>(lista, HttpStatus.OK);
     }
 
-    @GetMapping("/medico/citados")
-    public ResponseEntity<List<Cita>> listaCitados(
-            @RequestParam("idMedico") Integer idMedico, @RequestParam("numeroDiaSemana")Integer numeroDiaSemana) throws ResourceNotFoundException {
-        List<Cita> lista = new ArrayList<>();
-
-        lista = iCitaService.listaCitados(idMedico, numeroDiaSemana);
-        if (lista.isEmpty()) {
+    @GetMapping("/listaCitados")
+    public ResponseEntity<Set<CitaResponse>> listaCitados(
+            @RequestParam("idEmpresa") Integer idEmpresa, 
+            @RequestParam("numeroDocumento") String numeroDocumento,
+            @RequestParam("numeroDiaSemana")Integer numeroDiaSemana) throws ResourceNotFoundException {
+       System.out.println("idEmpresa "+  idEmpresa); 
+       System.out.println("numeroDocumento "+  numeroDocumento); 
+       System.out.println("numeroDiaSemana "+  numeroDiaSemana); 
+           
+    	var lista = iCitaService.listaCitados(idEmpresa,numeroDocumento, numeroDiaSemana);
+    	System.out.println("lista "+  lista); 
+    	if (lista.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-        return new ResponseEntity<List<Cita>>(lista, HttpStatus.OK);
+        return new ResponseEntity<>(lista, HttpStatus.OK);
     }
     
 
     @GetMapping("/medico")
-    public ResponseEntity<List<Cita>> listarCitas(
+    public ResponseEntity<Set<CitaResponse>> listarCitas(
             @RequestParam("idProgramacionDetalle") Integer idProgramacionDetalle) throws ResourceNotFoundException {
-        List<Cita> lista = new ArrayList<>();
-
-        lista = iCitaService.listarCitas(idProgramacionDetalle);
+        var lista = iCitaService.listarCitas(idProgramacionDetalle);
         if (lista.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-        return new ResponseEntity<List<Cita>>(lista, HttpStatus.OK);
+        return new ResponseEntity<>(lista, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Cita> listarPorId(@PathVariable("id") Integer id) throws ResourceNotFoundException {
-        Cita obj = iCitaService.listarPorId(id);
+    public ResponseEntity<CitaResponse> listarPorId(@PathVariable("id") Integer id) throws ResourceNotFoundException {
+        var obj = iCitaService.listarPorId(id);
 
         if (obj.getIdCita() == null) {
             throw new ResourceNotFoundException("Id no encontrado " + id);
         }
 
-        return new ResponseEntity<Cita>(obj, HttpStatus.OK);
+        return new ResponseEntity<>(obj, HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<Cita> registrar(@Valid @RequestBody CitaRequest citaRequest) throws ResourceNotFoundException {
-
-        Cita obj = iCitaService.registrar(Cita.builder()
-                .horario(Horario.builder().idHorario(citaRequest.getIdHorario()).build())
-                .programacionDetalle(ProgramacionDetalle.builder()
-                        .idProgramacionDetalle(citaRequest.getIdProgramacionDetalle()).build())
-                .cliente(Cliente.builder().idCliente(citaRequest.getIdCliente()).build())
-                .atendido(citaRequest.getAtendido())
-                .build());
+    public ResponseEntity<CitaResponse> registrar(@Valid @RequestBody CitaRequest citaRequest) throws ResourceNotFoundException {
+    	var obj =iCitaService.registrar(citaRequest);
+       
         return new ResponseEntity<>(obj, HttpStatus.CREATED);
     }
 
     @PutMapping
-    public ResponseEntity<Cita> modificar(@Valid @RequestBody CitaRequest citaRequest) throws ResourceNotFoundException {
-        Cita objRequest = Cita.builder()
-                .idCita(citaRequest.getIdCita())
-                .cliente(Cliente.builder().idCliente(citaRequest.getIdCliente()).build())
-                .programacionDetalle(ProgramacionDetalle.builder().idProgramacionDetalle(citaRequest.getIdProgramacionDetalle()).build())
-                .horario(Horario.builder().idHorario(citaRequest.getIdHorario()).build())
-                .atendido(citaRequest.getAtendido())
-                .informe(citaRequest.getInforme())
-                .build();
-        Cita obj = iCitaService.modificar(objRequest);
-        return new ResponseEntity<Cita>(obj, HttpStatus.OK);
+    public ResponseEntity<CitaResponse> modificar(@Valid @RequestBody CitaRequest citaRequest) throws ResourceNotFoundException {
+       
+        var obj = iCitaService.modificar(citaRequest);
+        return new ResponseEntity<>(obj, HttpStatus.OK);
     }
 
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> eliminar(@PathVariable("id") Integer id) throws Exception {
-        Cita obj = iCitaService.listarPorId(id);
+        var obj = iCitaService.listarPorId(id);
 
         if (obj == null) {
             throw new ResourceNotFoundException("ID NO ENCONTRADO " + id);
@@ -123,34 +107,37 @@ public class CitaController {
     }
     
     @GetMapping("/historial/pageable")
-    public ResponseEntity<Page<Cita>> listarPageableHistorico(@RequestParam("idCliente") Integer idCliente, @RequestParam(defaultValue = "0") int page,
+    public ResponseEntity<Page<CitaResponse>> listarPageableHistorico(
+    		@RequestParam("numeroDocumento") String numeroDocumento, 
+    		@RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "3") int size) throws ResourceNotFoundException {
+    	System.out.println("numeroDocumento  "+ numeroDocumento);
         System.out.println("page  "+ page);
         System.out.println("size  "+ size);
 
-        Pageable paging = PageRequest.of(page, size, Sort.by("idCita").descending());
-        Page<Cita> paginas  = iCitaService.listaHistorialCitaCliente(idCliente, paging);
+        Pageable paging = PageRequest.of(page, size);
+        Page<CitaResponse> paginas  = iCitaService.listaHistorialCitaCliente(numeroDocumento, paging);
         System.out.println("listarPageableCitas paginas "+ paginas);
 
         if (paginas.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-        return new ResponseEntity<Page<Cita>>(paginas, HttpStatus.OK);
+        return new ResponseEntity<>(paginas, HttpStatus.OK);
     }
 
     @GetMapping("/pageable")
-    public ResponseEntity<Page<Cita>> listarPageable(Pageable pageable, @RequestParam(defaultValue = "0") int page,
+    public ResponseEntity<Page<CitaResponse>> listarPageable(Pageable pageable, @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "3") int size) throws ResourceNotFoundException {
         System.out.println("page  "+ page);
         System.out.println("size  "+ size);
 
         
-        Page<Cita> paginas  = iCitaService.listarPageable( pageable);
+        Page<CitaResponse> paginas  = iCitaService.listarPageable( pageable);
         System.out.println("listarPageableCitas paginas "+ paginas);
 
         if (paginas.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-        return new ResponseEntity<Page<Cita>>(paginas, HttpStatus.OK);
+        return new ResponseEntity<>(paginas, HttpStatus.OK);
     }
 }
