@@ -1,8 +1,6 @@
 package com.gafahtec.consultorio.controller;
 
-import java.util.List;
-
-import javax.validation.Valid;
+import java.util.Set;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,11 +17,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.gafahtec.consultorio.dto.request.EmpresaRequest;
+import com.gafahtec.consultorio.dto.response.EmpresaResponse;
 import com.gafahtec.consultorio.exception.ResourceNotFoundException;
-import com.gafahtec.consultorio.model.Cliente;
-import com.gafahtec.consultorio.model.Empresa;
 import com.gafahtec.consultorio.service.IEmpresaService;
 
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 
 @RestController
@@ -34,40 +33,40 @@ public class EmpresaController {
 	private IEmpresaService iEmpresaService;
 	
 	@GetMapping
-	public ResponseEntity<List<Empresa>> listar() throws Exception{
-		List<Empresa> lista = iEmpresaService.listar();
+	public ResponseEntity<Set<EmpresaResponse>> listar() throws Exception{
+		var lista = iEmpresaService.listar();
 		if (lista.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-		return new ResponseEntity<List<Empresa>>(lista, HttpStatus.OK);
+		return new ResponseEntity<>(lista, HttpStatus.OK);
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<Empresa> listarPorId(@PathVariable("id") Integer id) throws Exception{
-		Empresa obj = iEmpresaService.listarPorId(id);
+	public ResponseEntity<EmpresaResponse> listarPorId(@PathVariable("id") Integer id) throws Exception{
+		var obj = iEmpresaService.listarPorId(id);
 		
 		if(obj.getIdEmpresa() == null) {
 		    throw new ResourceNotFoundException("Id no encontrado " + id);
 		}
 		
-		return new ResponseEntity<Empresa>(obj, HttpStatus.OK);
+		return new ResponseEntity<>(obj, HttpStatus.OK);
 	}
 	
 	@PostMapping
-	public ResponseEntity<Empresa> registrar(@Valid @RequestBody Empresa p) throws Exception{
-		Empresa obj = iEmpresaService.registrar(p);
+	public ResponseEntity<EmpresaResponse> registrar(@Valid @RequestBody EmpresaRequest empresaRequest) throws Exception{
+		var obj = iEmpresaService.registrar(empresaRequest);
 		 return new ResponseEntity<>(obj, HttpStatus.CREATED);
 	}
 	
 	@PutMapping
-	public ResponseEntity<Empresa> modificar(@Valid @RequestBody Empresa p) throws Exception{
-		Empresa obj = iEmpresaService.modificar(p);
-		return new ResponseEntity<Empresa>(obj, HttpStatus.OK);
+	public ResponseEntity<EmpresaResponse> modificar(@Valid @RequestBody EmpresaRequest empresaRequest) throws Exception{
+		var obj = iEmpresaService.modificar(empresaRequest);
+		return new ResponseEntity<>(obj, HttpStatus.OK);
 	}
 	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> eliminar(@PathVariable("id") Integer id) throws Exception{
-		Empresa obj = iEmpresaService.listarPorId(id);
+		var obj = iEmpresaService.listarPorId(id);
 		
 		if(obj.getIdEmpresa() == null) {
 			throw new ResourceNotFoundException("ID NO ENCONTRADO "+id);
@@ -78,12 +77,12 @@ public class EmpresaController {
 	}
 	
 	   @GetMapping("/pageable")
-	    public ResponseEntity<Page<Empresa>> listarPageable(@PageableDefault(sort = "nombre")Pageable pageable,@RequestParam(defaultValue = "0") int page,
+	    public ResponseEntity<Page<EmpresaResponse>> listarPageable(@PageableDefault(sort = "nombre")Pageable pageable,@RequestParam(defaultValue = "0") int page,
 	            @RequestParam(defaultValue = "5") int size) throws Exception{
-	        Page<Empresa> paginas = iEmpresaService.listarPageable(pageable);
+	        var paginas = iEmpresaService.listarPageable(pageable);
 	        if (paginas.isEmpty()) {
 	            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	        }
-	        return new ResponseEntity<Page<Empresa>>(paginas, HttpStatus.OK);
+	        return new ResponseEntity<>(paginas, HttpStatus.OK);
 	    }
 }
