@@ -1,10 +1,9 @@
 package com.gafahtec.consultorio.controller;
 
 import java.util.List;
+import java.util.Set;
 
-import javax.validation.Valid;
-
-import org.apache.commons.beanutils.BeanUtils;
+import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,11 +16,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.gafahtec.consultorio.dto.request.HistoriaClinicaRequest;
+import com.gafahtec.consultorio.dto.response.HistoriaClinicaResponse;
 import com.gafahtec.consultorio.exception.ResourceNotFoundException;
-import com.gafahtec.consultorio.model.Cliente;
-import com.gafahtec.consultorio.model.HistoriaClinica;
+import com.gafahtec.consultorio.model.consultorio.Cliente;
+import com.gafahtec.consultorio.model.consultorio.HistoriaClinica;
+import com.gafahtec.consultorio.service.IClienteService;
 import com.gafahtec.consultorio.service.IHistoriaClinicaService;
 
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 @RestController
 @RequestMapping("/historiasClinicas")
@@ -29,55 +31,51 @@ import lombok.AllArgsConstructor;
 public class HistoriaClinicaController {
 
 	private IHistoriaClinicaService iHistoriaClinicaService;
-	
+
 	@GetMapping
-	public ResponseEntity<List<HistoriaClinica>> listar() throws Exception{
-		List<HistoriaClinica> lista = iHistoriaClinicaService.listar();
+	public ResponseEntity<Set<HistoriaClinicaResponse>> listar() throws Exception{
+		var lista = iHistoriaClinicaService.listar();
 		if (lista.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-		return new ResponseEntity<List<HistoriaClinica>>(lista, HttpStatus.OK);
+		return new ResponseEntity<>(lista, HttpStatus.OK);
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<HistoriaClinica> listarPorId(@PathVariable("id") Integer id) throws Exception{
-		HistoriaClinica obj = iHistoriaClinicaService.listarPorId(id);
+	public ResponseEntity<HistoriaClinicaResponse> listarPorId(@PathVariable("id") Integer id) throws Exception{
+		var obj = iHistoriaClinicaService.listarPorId(id);
 		
 		if(obj.getIdHistoriaClinica() == null) {
 		    throw new ResourceNotFoundException("Id no encontrado " + id);
 		}
 		
-		return new ResponseEntity<HistoriaClinica>(obj, HttpStatus.OK);
+		return new ResponseEntity<>(obj, HttpStatus.OK);
 	}
 	
 	@PostMapping
-	public ResponseEntity<HistoriaClinica> registrar(@Valid @RequestBody HistoriaClinicaRequest historiaClinicaRequest) throws Exception{
-		HistoriaClinica historiaClinica = new HistoriaClinica();
-		BeanUtils.copyProperties(historiaClinica, historiaClinicaRequest);
-		historiaClinica.setCliente(Cliente.builder().idCliente(historiaClinicaRequest.getIdCliente()).build());
-		HistoriaClinica obj = iHistoriaClinicaService.registrar(historiaClinica);
+	public ResponseEntity<HistoriaClinicaResponse> registrar(@Valid @RequestBody HistoriaClinicaRequest historiaClinicaRequest) throws Exception{
+
+		var obj = iHistoriaClinicaService.registrar(historiaClinicaRequest);
 		 return new ResponseEntity<>(obj, HttpStatus.CREATED);
 	}
 	
 	@PutMapping
-	public ResponseEntity<HistoriaClinica> modificar(@Valid @RequestBody HistoriaClinicaRequest historiaClinicaRequest) throws Exception{
-		HistoriaClinica historiaClinica = new HistoriaClinica();
-		BeanUtils.copyProperties(historiaClinica, historiaClinicaRequest);
-		historiaClinica.setCliente(Cliente.builder().idCliente(historiaClinicaRequest.getIdCliente()).build());
-		HistoriaClinica obj = iHistoriaClinicaService.modificar(historiaClinica);
-		return new ResponseEntity<HistoriaClinica>(obj, HttpStatus.OK);
+	public ResponseEntity<HistoriaClinicaResponse> modificar(@Valid @RequestBody HistoriaClinicaRequest historiaClinicaRequest) throws Exception{
+
+		var obj = iHistoriaClinicaService.modificar(historiaClinicaRequest);
+		return new ResponseEntity<>(obj, HttpStatus.OK);
 	}
 	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> eliminar(@PathVariable("id") Integer id) throws Exception{
-		HistoriaClinica obj = iHistoriaClinicaService.listarPorId(id);
+		var obj = iHistoriaClinicaService.listarPorId(id);
 		
 		if(obj.getIdHistoriaClinica() == null) {
 			throw new ResourceNotFoundException("ID NO ENCONTRADO "+id);
 		}
 			
 		iHistoriaClinicaService.eliminar(id);
-		return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
+		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
 	
 //  @GetMapping("/pageable")
