@@ -1,8 +1,7 @@
 package com.gafahtec.consultorio.controller;
 
 import java.util.List;
-
-import javax.validation.Valid;
+import java.util.Set;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,10 +18,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.gafahtec.consultorio.dto.request.RolRequest;
+import com.gafahtec.consultorio.dto.response.RolResponse;
 import com.gafahtec.consultorio.exception.ResourceNotFoundException;
-import com.gafahtec.consultorio.model.Rol;
+import com.gafahtec.consultorio.model.auth.Rol;
 import com.gafahtec.consultorio.service.IRolService;
 
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 
 @RestController
@@ -33,40 +35,40 @@ public class RolController {
 	private IRolService iRolService;
 	
 	@GetMapping
-	public ResponseEntity<List<Rol>> listar() throws Exception{
-		List<Rol> lista = iRolService.listar();
+	public ResponseEntity<Set<RolResponse>> listar() throws Exception{
+		var lista = iRolService.listar();
 		if (lista.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
-		return new ResponseEntity<List<Rol>>(lista, HttpStatus.OK);
+		return new ResponseEntity<>(lista, HttpStatus.OK);
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<Rol> listarPorId(@PathVariable("id") Integer id) throws Exception{
-		Rol obj = iRolService.listarPorId(id);
+	public ResponseEntity<RolResponse> listarPorId(@PathVariable("id") Integer id) throws Exception{
+		var obj = iRolService.listarPorId(id);
 		
 		if(obj.getIdRol() == null) {
 		    throw new ResourceNotFoundException("Id no encontrado " + id);
 		}
 		
-		return new ResponseEntity<Rol>(obj, HttpStatus.OK);
+		return new ResponseEntity<>(obj, HttpStatus.OK);
 	}
 	
 	@PostMapping
-	public ResponseEntity<Rol> registrar(@Valid @RequestBody Rol p) throws Exception{
-		Rol obj = iRolService.registrar(p);
+	public ResponseEntity<RolResponse> registrar(@Valid @RequestBody RolRequest p) throws Exception{
+		var obj = iRolService.registrar(p);
 		 return new ResponseEntity<>(obj, HttpStatus.CREATED);
 	}
 	
 	@PutMapping
-	public ResponseEntity<Rol> modificar(@Valid @RequestBody Rol p) throws Exception{
-		Rol obj = iRolService.modificar(p);
-		return new ResponseEntity<Rol>(obj, HttpStatus.OK);
+	public ResponseEntity<RolResponse> modificar(@Valid @RequestBody RolRequest p) throws Exception{
+		var obj = iRolService.modificar(p);
+		return new ResponseEntity<>(obj, HttpStatus.OK);
 	}
 	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> eliminar(@PathVariable("id") Integer id) throws Exception{
-		Rol obj = iRolService.listarPorId(id);
+		var obj = iRolService.listarPorId(id);
 		
 		if(obj.getIdRol() == null) {
 			throw new ResourceNotFoundException("ID NO ENCONTRADO "+id);
@@ -77,12 +79,12 @@ public class RolController {
 	}
 	
 	  @GetMapping("/pageable")
-      public ResponseEntity<Page<Rol>> listarPageable(@PageableDefault(sort = "nombre")Pageable pageable,@RequestParam(defaultValue = "0") int page,
+      public ResponseEntity<Page<RolResponse>> listarPageable(@PageableDefault(sort = "nombre")Pageable pageable,@RequestParam(defaultValue = "0") int page,
               @RequestParam(defaultValue = "5") int size) throws Exception{
-          Page<Rol> paginas = iRolService.listarPageable(pageable);
+          var paginas = iRolService.listarPageable(pageable);
           if (paginas.isEmpty()) {
               return new ResponseEntity<>(HttpStatus.NO_CONTENT);
           }
-          return new ResponseEntity<Page<Rol>>(paginas, HttpStatus.OK);
+          return new ResponseEntity<>(paginas, HttpStatus.OK);
       }
 }
