@@ -2,8 +2,9 @@ package com.gafahtec.consultorio.dto.request;
 
 import java.time.LocalDateTime;
 
-import com.gafahtec.consultorio.model.HistoriaClinica;
+import com.gafahtec.consultorio.model.consultorio.HistoriaClinica;
 
+import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Getter;
@@ -22,10 +23,14 @@ import lombok.ToString;
 @ToString
 public class ClienteRequest {
     private Integer idCliente;
+    @NotEmpty(message = "Campo nombres no puede estar vacio")
     private String nombres;
+    @NotEmpty(message = "Campo apellido paterno no puede estar vacio")
     private String apellidoPaterno;
+    @NotEmpty(message = "Campo apellido materno no puede estar vacio")
     private String apellidoMaterno;
     private String tipoDocumento;
+    @NotEmpty(message = "Numero de documento no puede estar vacio")
     private String numeroDocumento;
     private String direccion;
     private String telefono;
