@@ -3,23 +3,15 @@ package com.gafahtec.consultorio.dto.request;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-@Setter
 
-@Getter
-
-@AllArgsConstructor
-
-@NoArgsConstructor
-
-@ToString
+@Data
 @Builder
+@AllArgsConstructor
+@NoArgsConstructor
 public class UsuarioRequest {
+
 	private String email;
-	private String password;
 	private String nombres;
 	private String numeroDocumento;
 	private String apellidoPaterno;
@@ -27,5 +19,5 @@ public class UsuarioRequest {
 	private Integer idRol;
 	private Integer idEmpresa;
 	private Integer idUsuario;
-	private Integer idTipoEmpleado;
+	private Integer idEmpleado;
 }
