@@ -2,8 +2,8 @@ package com.gafahtec.consultorio.repository;
 
 import org.springframework.stereotype.Repository;
 
-import com.gafahtec.consultorio.model.Cliente;
+import com.gafahtec.consultorio.model.consultorio.Cliente;
 
 @Repository
-public interface IClienteRepository extends IGenericRepository<Cliente,Integer>{
+public interface IClienteRepository extends IGenericRepository<Cliente,String>{
 }
