@@ -2,7 +2,7 @@ package com.gafahtec.consultorio.repository;
 
 import org.springframework.stereotype.Repository;
 
-import com.gafahtec.consultorio.model.HistoriaClinica;
+import com.gafahtec.consultorio.model.consultorio.HistoriaClinica;
 
 @Repository
 public interface IHistoriaClinicaRepository extends IGenericRepository<HistoriaClinica,Integer>{
