@@ -1,7 +1,8 @@
 package com.gafahtec.consultorio.repository;
 
-import com.gafahtec.consultorio.model.Horario;
 import org.springframework.stereotype.Repository;
+
+import com.gafahtec.consultorio.model.consultorio.Horario;
 
 @Repository
 public interface IHorarioRepository extends IGenericRepository<Horario,Integer>{
