@@ -1,6 +1,7 @@
 package com.gafahtec.consultorio.repository;
 
 import java.util.List;
+import java.util.Set;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -8,15 +9,18 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.gafahtec.consultorio.model.Programacion;
+import com.gafahtec.consultorio.model.consultorio.Programacion;
 
 @Repository
 public interface IProgramacionRepository extends IGenericRepository<Programacion,Integer>{
-
+    Set<Programacion> findByActivo(Boolean activo);
+    
     List<Programacion> findByRango(String rango);
     
-    List<Programacion> findByEstado(Boolean estado);
-    @Query("Select p from Programacion p where p.idEmpresa = :idEmpresa " )
+
+    
+    @Query("Select p from Programacion p  " )
+//    @Query("Select p from Programacion p where p.idEmpresa = :idEmpresa " )
     Page<Programacion> listarProgramacionPageable(@Param("idEmpresa") Integer idEmpresa, Pageable pageable);
     
     
