@@ -1,6 +1,6 @@
 package com.gafahtec.consultorio.repository;
 
-import com.gafahtec.consultorio.model.Rol;
+import com.gafahtec.consultorio.model.auth.Rol;
 
 public interface IRolRepository extends IGenericRepository<Rol,Integer>{
 	
