@@ -1,9 +1,11 @@
 package com.gafahtec.consultorio.repository;
 
-import com.gafahtec.consultorio.model.Usuario;
+import java.util.Optional;
+
+import com.gafahtec.consultorio.model.auth.Usuario;
 
 public interface IUsuarioRepository extends IGenericRepository<Usuario,Integer>{
 
-	 Usuario findByEmail(String email);
+	Optional<Usuario> findByEmail(String email);
 
 }
