@@ -1,7 +1,8 @@
 package com.gafahtec.consultorio.service;
 
-import com.gafahtec.consultorio.model.HistoriaClinica;
+import com.gafahtec.consultorio.dto.request.HistoriaClinicaRequest;
+import com.gafahtec.consultorio.dto.response.HistoriaClinicaResponse;
 
 
-public interface IHistoriaClinicaService extends ICRUD<HistoriaClinica,Integer>{
+public interface IHistoriaClinicaService extends ICRUD<HistoriaClinicaRequest, HistoriaClinicaResponse,Integer>{
 }
