@@ -3,9 +3,10 @@ package com.gafahtec.consultorio.service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import com.gafahtec.consultorio.model.Horario;
+import com.gafahtec.consultorio.dto.request.HorarioRequest;
+import com.gafahtec.consultorio.dto.response.HorarioResponse;
 
 
-public interface IHorarioService extends ICRUD<Horario,Integer>{
-    Page<Horario> listarPageable(Pageable pageable);
+public interface IHorarioService extends ICRUD<HorarioRequest,HorarioResponse,Integer>{
+    Page<HorarioResponse> listarPageable(Pageable pageable);
 }
