@@ -1,188 +1,233 @@
 package com.gafahtec.consultorio.service.impl;
 
-import java.util.ArrayList;
+import java.io.IOException;
 import java.util.HashSet;
-import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
+import com.gafahtec.consultorio.repository.*;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.gafahtec.consultorio.config.JwtService;
+import com.gafahtec.consultorio.dto.request.AuthenticationRequest;
+import com.gafahtec.consultorio.dto.request.AuthenticationResponse;
 import com.gafahtec.consultorio.dto.request.UsuarioRequest;
-import com.gafahtec.consultorio.model.Empleado;
-import com.gafahtec.consultorio.model.Empresa;
-import com.gafahtec.consultorio.model.Persona;
-import com.gafahtec.consultorio.model.Rol;
-import com.gafahtec.consultorio.model.TipoEmpleado;
-import com.gafahtec.consultorio.model.Usuario;
-import com.gafahtec.consultorio.repository.IEmpleadoRepository;
-import com.gafahtec.consultorio.repository.IGenericRepository;
-import com.gafahtec.consultorio.repository.IPersonaRepository;
-import com.gafahtec.consultorio.repository.IUsuarioRepository;
-import com.gafahtec.consultorio.service.UsuarioService;
+import com.gafahtec.consultorio.dto.request.UsuarioResponse;
+import com.gafahtec.consultorio.enums.TokenType;
+import com.gafahtec.consultorio.model.auth.Empleado;
+import com.gafahtec.consultorio.model.auth.Rol;
+import com.gafahtec.consultorio.model.auth.Token;
+import com.gafahtec.consultorio.model.auth.Usuario;
+import com.gafahtec.consultorio.service.IUsuarioService;
 import com.gafahtec.consultorio.util.Constants;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
-
-@AllArgsConstructor
 @Service
-
+@Transactional
+@AllArgsConstructor
 @Log4j2
-public class UsuarioServiceImpl extends CRUDImpl<Usuario, Integer> implements UsuarioService, UserDetailsService {
+public class UsuarioServiceImpl  implements IUsuarioService {
+	private final IUsuarioRepository iUsuarioRepository;
+	private final ITokenRepository iTokenRepository;
+	private final PasswordEncoder passwordEncoder;
+	private final JwtService jwtService;
+	private final AuthenticationManager authenticationManager;
+	private final IEmpresaRepository iEmpresaRepository;
+	private final IEmpleadoRepository iEmpleadoRepository;
+	private final IRolRepository iRolRepository;
+	
 
-    private IUsuarioRepository repo;
-    private IEmpleadoRepository iEmpleadoRepository;
+	public Usuario register(UsuarioRequest usuarioRequest) {
 
-    private IPersonaRepository iPersonaRepository;
-//    private IRolRepository rolRepo;
+		var usuarioRegistrado = iUsuarioRepository.findByEmail(usuarioRequest.getEmail());
 
-    private final PasswordEncoder passwordEncoder;
+		if (usuarioRegistrado.isEmpty()) {
+			var empresa = iEmpresaRepository
+					.findById(usuarioRequest.getIdEmpresa())
+					.orElseThrow(() -> new EntityNotFoundException("Empresa no encontrada con ID: " + usuarioRequest.getIdEmpresa()));
 
-//	private AuthTokenRepository authTokenRepository;
+			var empleado = Empleado.builder()
+//					.idEmpleado(usuarioRequest.getIdEmpleado())
+					.activo(Constants.ACTIVO)
+					.nombres(usuarioRequest.getNombres())
+					.apellidoPaterno(usuarioRequest.getApellidoPaterno())
+					.apellidoMaterno(usuarioRequest.getApellidoMaterno())
+					.numeroDocumento(usuarioRequest.getNumeroDocumento())
+					.empresa(empresa)
+					.build();
 
-    @Override
-    protected IGenericRepository<Usuario, Integer> getRepo() {
+			var savedEmpleado = iEmpleadoRepository.save(empleado);
+	            
+			Set<Rol> hashRoles = new HashSet<Rol>();
+			 hashRoles.add(iRolRepository.findById(usuarioRequest.getIdRol()).get());
 
-        return repo;
-    }
+			var user = Usuario.builder()
+					
+					.email(usuarioRequest.getEmail())
+					.password(passwordEncoder.encode("123456"))
+					.empleado(savedEmpleado)
+					.roles(hashRoles).build();
+			
+			var savedUser = iUsuarioRepository.save(user);
+			var jwtToken = jwtService.generateToken(user);
 
-    @Transactional
-    @Override
-    public Usuario registrarUsuarioEmpleado(UsuarioRequest usuarioRequest) {
+			var refreshToken = jwtService.generateRefreshToken(user);
+			saveUserToken(savedUser, jwtToken);
 
-        var usuarioRegistrado = repo.findByEmail(usuarioRequest.getEmail());
+			return savedUser;
+//			return AuthenticationResponse.builder().accessToken(jwtToken).refreshToken(refreshToken).build();
+		}
+			log.info("Usuario ya se encuentra regostrado ");
+			return null;
 
-        if (usuarioRegistrado == null) {
+	}
 
-            var persona = Persona.builder().nombres(usuarioRequest.getNombres())
-                    .apellidoPaterno(usuarioRequest.getApellidoPaterno())
-                    .apellidoMaterno(usuarioRequest.getApellidoMaterno())
-                    .numeroDocumento(usuarioRequest.getNumeroDocumento())
-                    .build();
+	public AuthenticationResponse authenticate(AuthenticationRequest request) {
+		authenticationManager
+				.authenticate(new UsernamePasswordAuthenticationToken(request.getEmail(), request.getPassword()));
+		var user = iUsuarioRepository.findByEmail(request.getEmail()).orElseThrow();
+		var jwtToken = jwtService.generateToken(user);
+		var refreshToken = jwtService.generateRefreshToken(user);
+		revokeAllUserTokens(user);
+		saveUserToken(user, jwtToken);
+		return AuthenticationResponse.builder().accessToken(jwtToken).refreshToken(refreshToken).build();
+	}
 
-            var empresa = Empresa.builder().idEmpresa(usuarioRequest.getIdEmpresa()).build();
-            var tipoEmpleado = TipoEmpleado.builder().idTipoEmpleado(usuarioRequest.getIdTipoEmpleado()).build();
-            var savedPersona = iPersonaRepository.save(persona);
+	private void saveUserToken(Usuario user, String jwtToken) {
+		var token = Token.builder().usuario(user).token(jwtToken).tokenType(TokenType.BEARER).expired(false)
+				.revoked(false).build();
+		iTokenRepository.save(token);
+	}
 
-            log.info("savedPersona grabado " + savedPersona);
+	private void revokeAllUserTokens(Usuario user) {
+		var validUserTokens = iTokenRepository.findAllValidTokenByUser(user.getIdUsuario());
+		if (validUserTokens.isEmpty())
+			return;
+		validUserTokens.forEach(token -> {
+			token.setExpired(true);
+			token.setRevoked(true);
+		});
+		iTokenRepository.saveAll(validUserTokens);
+	}
 
-            Set<Rol> hashRoles = new HashSet<Rol>();
-//			hashRoles.add(rolSearched);
-            hashRoles.add(Rol.builder().idRol(usuarioRequest.getIdRol()).build());
+	public void refreshToken(HttpServletRequest request, HttpServletResponse response) throws IOException {
+		final String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
+		final String refreshToken;
+		final String userEmail;
+		if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+			return;
+		}
+		refreshToken = authHeader.substring(7);
+		userEmail = jwtService.extractUsername(refreshToken);
+		if (userEmail != null) {
+			var user = this.iUsuarioRepository.findByEmail(userEmail).orElseThrow();
+			if (jwtService.isTokenValid(refreshToken, user)) {
+				var accessToken = jwtService.generateToken(user);
+				revokeAllUserTokens(user);
+				saveUserToken(user, accessToken);
+				var authResponse = AuthenticationResponse.builder().accessToken(accessToken).refreshToken(refreshToken)
+						.build();
+				new ObjectMapper().writeValue(response.getOutputStream(), authResponse);
+			}
+		}
+	}
+	
+	public  Optional<Usuario> findByEmail(String email){
+		
+		return iUsuarioRepository.findByEmail(email);
+	}
 
-            var empleado = Empleado.builder()
-                    .estado(Constants.ACTIVO)
-                    .persona(persona)
-                    .empresa(empresa)
-                    .tipoEmpleado(tipoEmpleado)
-                    .build();
-
-            var savedEmpleado = iEmpleadoRepository.save(empleado);
-
-            log.info("savedEmpleado grabado " + savedEmpleado);
-
-            var usuario = Usuario.builder().email(usuarioRequest.getEmail())
-                    .password(passwordEncoder.encode(usuarioRequest.getPassword()))
-                    .empleado(savedEmpleado)
-                    .roles(hashRoles).build();
-            log.info("usuario grabado " + usuario);
-            return repo.save(usuario);
-        } else {
-            return null;
-        }
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Usuario usuario = repo.findByEmail(email);
-
-        if (usuario == null) {
-            log.error("Usuario no se encuentra en base de datos");
-            throw new UsernameNotFoundException("Usuario no se encuentra en base de datos");
-        } else {
-            log.info("Usuario encontrado en  base de datos: {}", email);
-        }
-
-        List<GrantedAuthority> authorities = new ArrayList<>();
-
-        usuario.getRoles().forEach(rol -> {
-            authorities.add(new SimpleGrantedAuthority(rol.getNombre()));
-
-        });
-
-        return new User(usuario.getEmail(), usuario.getPassword(), authorities);
-    }
-
-    @Override
-    public Usuario getUsuario(String email) {
-        return repo.findByEmail(email);
-
-    }
 
     @Override
     public Page<Usuario> listarPageable(Pageable pageable) {
 
-        return repo.findAll(pageable);
+        return iUsuarioRepository.findAll(pageable);
     }
-
     @Override
     public Usuario getUsuarioPorId(Integer id) {
 
-        return repo.findById(id).get();
+        return iUsuarioRepository.findById(id).get();
     }
-
-    @Transactional
+    
+    
     @Override
-    public Usuario modificarUsuarioEmpleado(UsuarioRequest usuarioRequest) {
+    public Usuario modificarUsuario(UsuarioRequest usuarioRequest) {
 
-        var persona = Persona.builder().nombres(usuarioRequest.getNombres())
-                .apellidoPaterno(usuarioRequest.getApellidoPaterno())
-                .apellidoMaterno(usuarioRequest.getApellidoMaterno())
-                .numeroDocumento(usuarioRequest.getNumeroDocumento())
-                .build();
-
-        var empresa = Empresa.builder().idEmpresa(usuarioRequest.getIdEmpresa()).build();
-
-        var tipoEmpleado = TipoEmpleado.builder().idTipoEmpleado(usuarioRequest.getIdTipoEmpleado()).build();
-
-        var savedPersona = iPersonaRepository.save(persona);
-
-        log.info("savedPersona grabado " + savedPersona);
+		var empresa = iEmpresaRepository
+				.findById(usuarioRequest.getIdEmpresa())
+				.orElseThrow(() -> new EntityNotFoundException("Empresa no encontrada con ID: " + usuarioRequest.getIdEmpresa()));
 
         Set<Rol> hashRoles = new HashSet<Rol>();
-//          hashRoles.add(rolSearched);
-        hashRoles.add(Rol.builder().idRol(usuarioRequest.getIdRol()).build());
+        hashRoles.add(iRolRepository.findById(usuarioRequest.getIdRol()).get());
 
         var empleado = Empleado.builder()
-                .estado(Constants.ACTIVO)
-                .persona(persona)
+        		.idEmpleado(usuarioRequest.getIdEmpleado())
+                .activo(Constants.ACTIVO)
+				.nombres(usuarioRequest.getNombres())
+				.apellidoPaterno(usuarioRequest.getApellidoPaterno())
+				.apellidoMaterno(usuarioRequest.getApellidoMaterno())
+				.numeroDocumento(usuarioRequest.getNumeroDocumento())
                 .empresa(empresa)
-                .tipoEmpleado(tipoEmpleado)
                 .build();
 
-        var savedEmpleado = iEmpleadoRepository.save(empleado);
-
-        log.info("savedEmpleado grabado " + savedEmpleado);
-
-        var usuario = Usuario.builder().email(usuarioRequest.getEmail())
-                .password(passwordEncoder.encode(usuarioRequest.getPassword()))
-                .empleado(savedEmpleado)
-                .idUsuario(usuarioRequest.getIdUsuario())
-                .roles(hashRoles).build();
-        log.info("usuario grabado " + usuario);
-        return modificar(usuario);
+		iEmpleadoRepository.save(empleado);
+       
+        var usuario = iUsuarioRepository.findByEmail(usuarioRequest.getEmail()).get();
+		usuario.setEmpleado(empleado);
+        usuario.setRoles(hashRoles);
+        var userUpdate = iUsuarioRepository.save(usuario);
+        log.info("usuario grabado " + userUpdate);
+        return userUpdate;
 
     }
+
+	@Override
+	public UsuarioResponse registrar(UsuarioRequest request) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public UsuarioResponse modificar(UsuarioRequest request) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Set<UsuarioResponse> listar() {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public UsuarioResponse listarPorId(Integer id) {
+		return entityToResponse(iUsuarioRepository.findById(id).get());
+	}
+
+	@Override
+	public void eliminar(Integer id) {
+		// TODO Auto-generated method stub
+		
+	}
+
+	
+	private UsuarioResponse entityToResponse(Usuario entity) {
+		var response = new UsuarioResponse();
+		
+		return response;
+	}
 
 }
