@@ -4,5 +4,6 @@ public class Constants {
 
     public static final Boolean ACTIVO = true;
     public static final Boolean INACTIVO = false;
-
+    public static final Boolean ATENDIDO = true;
+    public static final Boolean DESATENDIDO = false;
 }
